Break down the CompositionRoot into smaller files
