+++ Phoenix/CompositionRoot/CompositionRoot+DemoAppGenerator.swift
@@ -1,0 +1,12 @@
+import DemoAppGenerator
+import DemoAppGeneratorContract
+import Factory
+import Foundation
+
+extension Container {
+    static let demoAppGenerator = Factory(Container.shared) {
+        DemoAppGenerator(
+            fileManager: FileManager.default
+        ) as DemoAppGeneratorProtocol
+    }
+}
