+++ Phoenix/CompositionRoot/CompositionRoot+DocumentCoder.swift
@@ -1,0 +1,15 @@
+import DocumentCoder
+import DocumentCoderContract
+import Factory
+
+extension Container {
+    static let phoenixDocumentFileWrappersDecoder = Factory(Container.shared) {
+        PhoenixDocumentFileWrappersDecoder() as PhoenixDocumentFileWrappersDecoderProtocol
+    }
+
+    static let phoenixDocumentFileWrapperEncoder = Factory(Container.shared) {
+        PhoenixDocumentFileWrapperEncoder(
+            currentAppVersionStringProvider: currentAppVersionStringProvider()
+        ) as PhoenixDocumentFileWrapperEncoderProtocol
+    }
+}
