+++ Phoenix/CompositionRoot/CompositionRoot+PackageGenerator.swift
@@ -1,0 +1,12 @@
+import Factory
+import PackageGenerator
+import PackageGeneratorContract
+
+extension Container {
+    static let packageGenerator = Factory(Container.shared) {
+        PackageGenerator(
+            fileManager: .default,
+            packageStringProvider: packageStringProvider()
+        ) as PackageGeneratorProtocol
+    }
+}
