+++ Phoenix/CompositionRoot/CompositionRoot+PackageStringProvider.swift
@@ -1,0 +1,9 @@
+import Factory
+import PackageStringProvider
+import PackageStringProviderContract
+
+extension Container {
+    static let packageStringProvider = Factory(Container.shared) {
+        PackageStringProvider() as PackageStringProviderProtocol
+    }
+}
