+++ Phoenix/CompositionRoot/CompositionRoot+RelativeURLProvider.swift
@@ -1,0 +1,9 @@
+import Factory
+import RelativeURLProvider
+import RelativeURLProviderContract
+
+extension Container {
+    static let relativeURLProvider = Factory(Container.shared) {
+        RelativeURLProvider() as RelativeURLProviderProtocol
+    }
+}
